+++ examples/editable_table.rb
@@ -20,6 +20,14 @@ window('Editable animal sounds', 300, 200) {
 
       cell_rows data
       editable true
+      
+      on_changed do |row, type, row_data| # fires on all changes (even ones happening through data array)
+        puts "Row #{row} #{type}: #{row_data}"
+      end
+      
+      on_edited do |row, row_data| # only fires on direct table editing
+        puts "Row #{row} edited: #{row_data}"
+      end
     }
   }